requirements.txt
tinygrad
Pillow
matplotlib
pandas
safetensors
